adding base e2e checking for test
