+++ docs/components/SideDrawer/SideDrawer.stories.mdx
@@ -1,4 +1,5 @@
 import { Meta, ArgsTable } from "@storybook/addon-docs";
+import { SideDrawer } from "@jobber/components";
 
 <Meta title="Components/Overlays/SideDrawer" parameters={{ alpha: true }} />
