+++ packages/site/baseComponentLists.mjs
@@ -1,0 +1,131 @@
+/**
+ * List of components to build documentation for. There is a future in which this list is dynamically generated from an 'ls' command or similar.
+ * For now though, since not ALL the components are in the system, we're building manually.
+ */
+const components = [
+  "AnimatedPresence",
+  "AnimatedSwitcher",
+  "Autocomplete",
+  "Avatar",
+  "Banner",
+  "Box",
+  "Button",
+  "ButtonDismiss",
+  "Card",
+  "Checkbox",
+  "Chip",
+  "Chips",
+  "Combobox",
+  "ConfirmationModal",
+  "Content",
+  "Countdown",
+  "DataDump",
+  "DataList",
+  "DataTable",
+  "DatePicker",
+  "DescriptionList",
+  "Disclosure",
+  "Divider",
+  "Drawer",
+  "Emphasis",
+  "FeatureSwitch",
+  "Flex",
+  "Form",
+  "FormatDate",
+  "FormatEmail",
+  "FormatFile",
+  "FormatRelativeDateTime",
+  "FormatTime",
+  "FormField",
+  "Gallery",
+  "Glimmer",
+  "Grid",
+  "Heading",
+  "Icon",
+  "InlineLabel",
+  "InputAvatar",
+  "InputDate",
+  "InputEmail",
+  "InputFile",
+  "InputGroup",
+  "InputNumber",
+  "InputPassword",
+  "InputPhoneNumber",
+  "InputText",
+  "InputTime",
+  "InputValidation",
+  "LightBox",
+  "Link",
+  "List",
+  "Markdown",
+  "Menu",
+  "Modal",
+  "MultiSelect",
+  "Page",
+  "Popover",
+  "ProgressBar",
+  "RadioGroup",
+  "RecurringSelect",
+  "SegmentedControl",
+  "Select",
+  "SideDrawer",
+  "Spinner",
+  "StatusIndicator",
+  "StatusLabel",
+  "Switch",
+  "Table",
+  "Tabs",
+  "Text",
+  "Toast",
+  "Tooltip",
+  "Typography",
+];
+
+const mobileComponents = [
+  "ActionLabel",
+  "Button",
+  "ButtonGroup",
+  "IconButton",
+  "Card",
+  "Form",
+  "InputDate",
+  "InputEmail",
+  "InputNumber",
+  "InputPassword",
+  "InputPressable",
+  "InputSearch",
+  "InputText",
+  "InputTime",
+  "FormatFile",
+  "Icon",
+  "ThumbnailList",
+  "ActionItem",
+  "ActionItemGroup",
+  "Content",
+  "Disclosure",
+  "Divider",
+  "Flex",
+  "TextList",
+  "Menu",
+  "ContentOverlay",
+  "FormField",
+  "InputFieldWrapper",
+  "BottomSheet",
+  "Checkbox",
+  "Chips",
+  "Select",
+  "Switch",
+  "ActivityIndicator",
+  "Banner",
+  "EmptyState",
+  "Glimmer",
+  "ProgressBar",
+  "StatusLabel",
+  "Toast",
+  "AutoLink",
+  "Heading",
+  "Text",
+  "Typography",
+];
+export const ListOfGeneratedMobileComponents = mobileComponents;
+export const ListOfGeneratedWebComponents = components;
